Create Windows AD script
This script grabs a list of all computers in AD and their OS build info, then exports it to a .csv spreadsheet. Pretty handy. I am documenting this script because it took a surprisingly long time to figure out how to do it. I was googling and thought "why hasn't anyone put this online?", then I thought I should be the change I want to see in the web.